Fixed builtin lib paths, crtend build flags on 3.2
3.2.x releases now have correct .eh_frame section in crtend.o and
crtendS.o, which current linkers want to see.  This is a backport from
a 2007 update.

The built in directories searched for "startfiles" - objects the gcc
driver passes to the linker first by default - are now fixed.

On 3.2.3, this reduces 4488 'make check-gcc' unexpected failures to 26
on Debian Stretch using the config in configure-opts.